+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - (Bugfix) Prevent Member Maintenance Error log
 - (Feature) ID ServerGroup
 - (Bugfix) Propagate Lifecycle Mount
+- (Feature) PVC Member Status info
 
 ## [1.2.19](https://github.com/arangodb/kube-arangodb/tree/1.2.19) (2022-10-05)
 - (Bugfix) Prevent changes when UID is wrong